Separate layers on movie save routine
